// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        BOM: {
            readonly id: number;
            readonly material_name: string;
            /**
             * Required Quantity
             * Format: decimal
             */
            required_qty: string;
            /** Format: decimal */
            actual_allocated_qty?: string;
            order: number;
            material: number;
        };
// ...
        PaginatedBOMList: {
            data: components["schemas"]["BOM"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
// ...
        PatchedBOM: {
            readonly id?: number;
            readonly material_name?: string;
            /**
             * Required Quantity
             * Format: decimal
             */
            required_qty?: string;
            /** Format: decimal */
            actual_allocated_qty?: string;
            order?: number;
            material?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}